cleanning old files
