+++ classifiers.py
@@ -1,47 +1,20 @@
-#import common as cm
-import numpy as np
-import matplotlib.pyplot as plt
-from matplotlib.colors import ListedColormap
 from sklearn.model_selection import train_test_split
-from sklearn.neural_network import MLPClassifier
 from sklearn.neighbors import KNeighborsClassifier
-from sklearn.svm import SVC
 from sklearn.tree import DecisionTreeClassifier
 from sklearn.ensemble import RandomForestClassifier, AdaBoostClassifier
-from sklearn.naive_bayes import GaussianNB
-from sklearn.discriminant_analysis import LinearDiscriminantAnalysis, QuadraticDiscriminantAnalysis
-from sklearn.linear_model import Perceptron
-# import xgboost as xgb
-
 
 names = [
-		#"Perceptron",
-		# "XGBreglinear",
-		# "XGBreglogistic",
 		"NearestNeighbors",
-		# "LinearSVM",
 		"DecisionTree",
 		"RandomForest",
-		"AdaBoost",
-		#"NeuralNet",
-		#"NaiveBayes", 
-		#"LDA",
-		#"QDA"
+		"AdaBoost"
 		]
 
 classifiers = [
-		#Perceptron(),
-		# xgb.XGBClassifier(objective='reg:linear'),
-		# xgb.XGBClassifier(objective='reg:logistic'),
 		KNeighborsClassifier(10),
-		# SVC(kernel="linear"),
 		DecisionTreeClassifier(max_depth=5),
 		RandomForestClassifier(max_depth=5, n_estimators=10, max_features=1),
-		AdaBoostClassifier(),
-		#MLPClassifier(verbose=False),
-		#GaussianNB(),
-		#LinearDiscriminantAnalysis(),
-		#QuadraticDiscriminantAnalysis()
+		AdaBoostClassifier()
     ]
