--- classify/transpose.py
@@ -1,4 +1,0 @@
-import common
-
-#common.transposeDataFile("colon-cancerT.csv", "colon-cancer.csv")
-common.transposeDataFile("f1.csv", "data-f1.csv")
