+++ data-factory/transpose.py
@@ -1,0 +1,7 @@
+import common as cm
+
+#cm.transposeDataFile("colon-cancerT.csv", "colon-cancer.csv")
+cm.transposeDataFile("f1.csv", "data-f1.csv")
+cm.transposeDataFile("f2.csv", "data-f2.csv")
+cm.transposeDataFile("f3.csv", "data-f3.csv")
+cm.transposeDataFile("f4.csv", "data-f4.csv")
