+++ test.py
@@ -4,14 +4,15 @@ import binDynamic as bd
 import binStatic as bs
 from pandas import read_csv, np
 import voteSelection as vs
+import util as ut
 
 def artificialTest():
 	#files = ['data-f1.csv','data-f2.csv','data-f3.csv','data-f4.csv','simpleCorrelations.csv','data-r50.csv','data2400-r20.csv']
 	#buenos = [[0,1,2,3,4,5,6],[0,1],[0,1],[0,1,3,2],[0,1,2,3,4,5],[0,1,2],[0,1,2]]
-	#files = ['data-f1.csv', 'data-f2.csv','data-f3.csv','data-f4.csv','data-r50.csv','data2400-r20.csv']
-	#buenos = [[0,1,2,3,4,5,6],[0,1],[0,1],[0,1,3,2],[0,1,2],[0,1,2]]
-	files = ['colon-cancer.csv']
-	buenos = [[0,1,2,3,4,5,6]]
+	files = ['data-f1.csv', 'data-f2.csv','data-f3.csv','data-f4.csv','data-r50.csv','data2400-r20.csv']
+	buenos = [[0,1,2,3,4,5,6],[0,1],[0,1],[0,1,3,2],[0,1,2],[0,1,2]]
+	#files = ['colon-cancer.csv']
+	#buenos = [[0,1,2,3,4,5,6]]
 	
 
 	i=0
@@ -30,15 +31,15 @@ def artificialTest():
 		ss = []
 		ss.append(bs.binStatic(data,0))
 		ss.append(bs.binStatic(data,1))
-		ss.append(bs.binStatic(data,2))
+		#ss.append(bs.binStatic(data,2))
 		#ss.append(bs.binStatic(data,3))
 		weight = (vs.sumMixedCorrelation(ss))
 		#print "static:",weight
 
 		#Dynamic search
-		#cr = []
+		cr = []
 		#binarySearchBinsX, y, method, split, useSteps, normalizeResult, Debug
-		#cr.append(bd.binarySearchBins(X,y,0,0,2))
+		cr.append(bd.binarySearchBins(X,y,0,0,2))
 		#cr.append(bd.binarySearchBins(X,y,1,0,2))
 		#cr.append(bd.binarySearchBins(X,y,2,0,2))
 		#cr = (vs.sumMixedCorrelation(cr))
@@ -47,7 +48,7 @@ def artificialTest():
 		#print bx
 		
 		########### Cuts ###########
-		rank = cuts.getOrderRank(weight)
+		rank = ut.getOrderRank(weight)
 
 		print cf.getBestClassifiers(X,y)
 		[X,y,cutpos] = cuts.greatestDiff(rank,weight,data)
